[Bronze V] Title: 이상한 기호, Time: 100 ms, Memory: 14212 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -5,11 +5,11 @@ public class Main {
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         StringTokenizer st = new StringTokenizer(br.readLine());
-        int A = Integer.parseInt(st.nextToken());
-        int B = Integer.parseInt(st.nextToken());
+        long A = Long.parseLong(st.nextToken());
+        long B = Long.parseLong(st.nextToken());
         System.out.print(cal(A,B));
     }
-    public static int cal(int a, int b) {
+    public static long cal(long a, long b) {
         return (a+b)*(a-b);
     }
 }
